add gate and update dob immediately

+++ app/Http/Controllers/ReviewController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Models\Review;
 use App\Models\Room;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class ReviewController extends Controller
 {
@@ -15,6 +16,10 @@ class ReviewController extends Controller
 
     public function store(Room $room, Request $request)
     {
+        if (!Gate::allows('user-reserved', $room)) {
+            abort(203);
+        }
+
         $this->validate($request, [
             'rating' => 'required',
             'review' => 'required',
@@ -31,6 +36,9 @@ class ReviewController extends Controller
 
     public function destroy(Review $review)
     {
+        if (!Gate::allows('user-reviews', $review)) {
+            abort(203);
+        }
         if ($review->reviewedBy(auth()->user())) {
             $review->delete();
         }
